# Add ByType logical rule