--- a/src/BenchmarkDotNet/Configs/BenchmarkLogicalGroupRule.cs
+++ b/src/BenchmarkDotNet/Configs/BenchmarkLogicalGroupRule.cs
@@ -2,6 +2,6 @@
 {
     public enum BenchmarkLogicalGroupRule
     {
-        ByMethod, ByJob, ByParams, ByCategory
+        ByMethod, ByJob, ByParams, ByCategory, ByType
     }
 }