--- a/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.Invalid_TwoJobBaselines.approved.txt
+++ b/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.Invalid_TwoJobBaselines.approved.txt
@@ -8,14 +8,14 @@ Frequency=2531248 Hz, Resolution=395.0620 ns, Timer=TSC
   Job2   : extra output line
 
 
- Method |  Job |     Mean |   Error |  StdDev | Ratio | RatioSD | Rank |                LogicalGroup | Baseline |
-------- |----- |---------:|--------:|--------:|------:|--------:|-----:|---------------------------- |--------- |
-    Foo | Job1 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 | Invalid_TwoJobBaselines.Foo |      Yes |
-    Foo | Job2 | 302.0 ns | 6.09 ns | 1.58 ns |  2.96 |    0.03 |    2 | Invalid_TwoJobBaselines.Foo |      Yes |
-        |      |          |         |         |       |         |      |                             |          |
-    Bar | Job1 | 202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 | Invalid_TwoJobBaselines.Bar |      Yes |
-    Bar | Job2 | 402.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 | Invalid_TwoJobBaselines.Bar |      Yes |
+ Method |  Job |     Mean |   Error |  StdDev | Ratio | RatioSD | Rank | LogicalGroup | Baseline |
+------- |----- |---------:|--------:|--------:|------:|--------:|-----:|------------- |--------- |
+    Foo | Job1 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |          Foo |      Yes |
+    Foo | Job2 | 302.0 ns | 6.09 ns | 1.58 ns |  2.96 |    0.03 |    2 |          Foo |      Yes |
+        |      |          |         |         |       |         |      |              |          |
+    Bar | Job1 | 202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |          Bar |      Yes |
+    Bar | Job2 | 402.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 |          Bar |      Yes |
 
 Errors: 2
-* Only 1 job in a group can have "Baseline = true" applied to it, group Invalid_TwoJobBaselines.Foo in class Invalid_TwoJobBaselines has 2
-* Only 1 job in a group can have "Baseline = true" applied to it, group Invalid_TwoJobBaselines.Bar in class Invalid_TwoJobBaselines has 2
+* Only 1 job in a group can have "Baseline = true" applied to it, group Foo in class Invalid_TwoJobBaselines has 2
+* Only 1 job in a group can have "Baseline = true" applied to it, group Bar in class Invalid_TwoJobBaselines has 2