--- a/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.JobBaseline_MethodsJobs.approved.txt
+++ b/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.JobBaseline_MethodsJobs.approved.txt
@@ -8,15 +8,15 @@ Frequency=2531248 Hz, Resolution=395.0620 ns, Timer=TSC
   Job2   : extra output line
 
 
- Method |  Job |     Mean |   Error |  StdDev | Ratio | RatioSD | Rank |                 LogicalGroup | Baseline |
-------- |----- |---------:|--------:|--------:|------:|--------:|-----:|----------------------------- |--------- |
-   Base | Job1 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 | JobBaseline_MethodsJobs.Base |      Yes |
-   Base | Job2 | 402.0 ns | 6.09 ns | 1.58 ns |  3.94 |    0.05 |    2 | JobBaseline_MethodsJobs.Base |       No |
-        |      |          |         |         |       |         |      |                              |          |
-    Foo | Job1 | 202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  JobBaseline_MethodsJobs.Foo |      Yes |
-    Foo | Job2 | 502.0 ns | 6.09 ns | 1.58 ns |  2.49 |    0.01 |    2 |  JobBaseline_MethodsJobs.Foo |       No |
-        |      |          |         |         |       |         |      |                              |          |
-    Bar | Job1 | 302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  JobBaseline_MethodsJobs.Bar |      Yes |
-    Bar | Job2 | 602.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 |  JobBaseline_MethodsJobs.Bar |       No |
+ Method |  Job |     Mean |   Error |  StdDev | Ratio | RatioSD | Rank | LogicalGroup | Baseline |
+------- |----- |---------:|--------:|--------:|------:|--------:|-----:|------------- |--------- |
+   Base | Job1 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |         Base |      Yes |
+   Base | Job2 | 402.0 ns | 6.09 ns | 1.58 ns |  3.94 |    0.05 |    2 |         Base |       No |
+        |      |          |         |         |       |         |      |              |          |
+    Foo | Job1 | 202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |          Foo |      Yes |
+    Foo | Job2 | 502.0 ns | 6.09 ns | 1.58 ns |  2.49 |    0.01 |    2 |          Foo |       No |
+        |      |          |         |         |       |         |      |              |          |
+    Bar | Job1 | 302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |          Bar |      Yes |
+    Bar | Job2 | 602.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 |          Bar |       No |
 
 Errors: 0