--- a/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.JobBaseline_MethodsParamsJobs.approved.txt
+++ b/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.JobBaseline_MethodsParamsJobs.approved.txt
@@ -8,24 +8,24 @@ Frequency=2531248 Hz, Resolution=395.0620 ns, Timer=TSC
   Job2   : extra output line
 
 
- Method |  Job | Param |       Mean |   Error |  StdDev | Ratio | RatioSD | Rank |                                  LogicalGroup | Baseline |
-------- |----- |------ |-----------:|--------:|--------:|------:|--------:|-----:|---------------------------------------------- |--------- |
-   Base | Job1 |     2 |   102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=2]-JobBaseline_MethodsParamsJobs.Base |      Yes | ^
-   Base | Job2 |     2 |   402.0 ns | 6.09 ns | 1.58 ns |  3.94 |    0.05 |    2 |  [Param=2]-JobBaseline_MethodsParamsJobs.Base |       No |
-        |      |       |            |         |         |       |         |      |                                               |          |
-    Foo | Job1 |     2 |   202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |   [Param=2]-JobBaseline_MethodsParamsJobs.Foo |      Yes |
-    Foo | Job2 |     2 |   502.0 ns | 6.09 ns | 1.58 ns |  2.49 |    0.01 |    2 |   [Param=2]-JobBaseline_MethodsParamsJobs.Foo |       No |
-        |      |       |            |         |         |       |         |      |                                               |          |
-    Bar | Job1 |     2 |   302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |   [Param=2]-JobBaseline_MethodsParamsJobs.Bar |      Yes |
-    Bar | Job2 |     2 |   602.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 |   [Param=2]-JobBaseline_MethodsParamsJobs.Bar |       No |
-        |      |       |            |         |         |       |         |      |                                               |          |
-   Base | Job1 |    10 |   702.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 | [Param=10]-JobBaseline_MethodsParamsJobs.Base |      Yes | ^
-   Base | Job2 |    10 | 1,002.0 ns | 6.09 ns | 1.58 ns |  1.43 |    0.00 |    2 | [Param=10]-JobBaseline_MethodsParamsJobs.Base |       No |
-        |      |       |            |         |         |       |         |      |                                               |          |
-    Foo | Job1 |    10 |   802.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=10]-JobBaseline_MethodsParamsJobs.Foo |      Yes |
-    Foo | Job2 |    10 | 1,102.0 ns | 6.09 ns | 1.58 ns |  1.37 |    0.00 |    2 |  [Param=10]-JobBaseline_MethodsParamsJobs.Foo |       No |
-        |      |       |            |         |         |       |         |      |                                               |          |
-    Bar | Job1 |    10 |   902.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=10]-JobBaseline_MethodsParamsJobs.Bar |      Yes |
-    Bar | Job2 |    10 | 1,202.0 ns | 6.09 ns | 1.58 ns |  1.33 |    0.00 |    2 |  [Param=10]-JobBaseline_MethodsParamsJobs.Bar |       No |
+ Method |  Job | Param |       Mean |   Error |  StdDev | Ratio | RatioSD | Rank |    LogicalGroup | Baseline |
+------- |----- |------ |-----------:|--------:|--------:|------:|--------:|-----:|---------------- |--------- |
+   Base | Job1 |     2 |   102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=2]-Base |      Yes | ^
+   Base | Job2 |     2 |   402.0 ns | 6.09 ns | 1.58 ns |  3.94 |    0.05 |    2 |  [Param=2]-Base |       No |
+        |      |       |            |         |         |       |         |      |                 |          |
+    Foo | Job1 |     2 |   202.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |   [Param=2]-Foo |      Yes |
+    Foo | Job2 |     2 |   502.0 ns | 6.09 ns | 1.58 ns |  2.49 |    0.01 |    2 |   [Param=2]-Foo |       No |
+        |      |       |            |         |         |       |         |      |                 |          |
+    Bar | Job1 |     2 |   302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |   [Param=2]-Bar |      Yes |
+    Bar | Job2 |     2 |   602.0 ns | 6.09 ns | 1.58 ns |  1.99 |    0.01 |    2 |   [Param=2]-Bar |       No |
+        |      |       |            |         |         |       |         |      |                 |          |
+   Base | Job1 |    10 |   702.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 | [Param=10]-Base |      Yes | ^
+   Base | Job2 |    10 | 1,002.0 ns | 6.09 ns | 1.58 ns |  1.43 |    0.00 |    2 | [Param=10]-Base |       No |
+        |      |       |            |         |         |       |         |      |                 |          |
+    Foo | Job1 |    10 |   802.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=10]-Foo |      Yes |
+    Foo | Job2 |    10 | 1,102.0 ns | 6.09 ns | 1.58 ns |  1.37 |    0.00 |    2 |  [Param=10]-Foo |       No |
+        |      |       |            |         |         |       |         |      |                 |          |
+    Bar | Job1 |    10 |   902.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |    1 |  [Param=10]-Bar |      Yes |
+    Bar | Job2 |    10 | 1,202.0 ns | 6.09 ns | 1.58 ns |  1.33 |    0.00 |    2 |  [Param=10]-Bar |       No |
 
 Errors: 0