--- a/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByCategory.approved.txt
+++ b/tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByCategory.approved.txt
@@ -1,0 +1,22 @@
+﻿=== NoBaseline.ByCategory ===
+
+BenchmarkDotNet=v0.10.x-mock, OS=Microsoft Windows NT 10.0.x.mock, VM=Hyper-V
+MockIntel Core i7-6700HQ CPU 2.60GHz (Max: 3.10GHz), 1 CPU, 8 logical and 4 physical cores
+Frequency=2531248 Hz, Resolution=395.0620 ns, Timer=TSC
+  [Host]     : Clr 4.0.x.mock, 64mock RyuJIT-v4.6.x.mock CONFIGURATION
+  DefaultJob : extra output line
+
+
+   Type | Method | Categories | Param |     Mean |   Error |  StdDev | LogicalGroup | Baseline |
+------- |------- |----------- |------ |---------:|--------:|--------:|------------- |--------- |
+ Bench1 |    Foo |          A |    10 | 102.0 ns | 6.09 ns | 1.58 ns |            A |       No | ^
+ Bench1 |    Bar |          A |    10 | 202.0 ns | 6.09 ns | 1.58 ns |            A |       No |
+ Bench1 |    Foo |          A |    20 | 302.0 ns | 6.09 ns | 1.58 ns |            A |       No | ^
+ Bench1 |    Bar |          A |    20 | 402.0 ns | 6.09 ns | 1.58 ns |            A |       No |
+        |        |            |       |          |         |         |              |          |
+ Bench2 |    Foo |          B |    10 | 102.0 ns | 6.09 ns | 1.58 ns |            B |       No | ^
+ Bench2 |    Bar |          B |    10 | 202.0 ns | 6.09 ns | 1.58 ns |            B |       No |
+ Bench2 |    Foo |          B |    20 | 302.0 ns | 6.09 ns | 1.58 ns |            B |       No | ^
+ Bench2 |    Bar |          B |    20 | 402.0 ns | 6.09 ns | 1.58 ns |            B |       No |
+
+Errors: 0