… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.NoBaseline_MethodsParamsJobs_GroupByAll.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterApprovalTests.GroupExporterTest.NoBaseline_MethodsParamsJobs_GroupByMethod.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByJob.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByMethod.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByParams.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.ByType.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.NoBaseline.Default.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.ByCategory.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.ByJob.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.ByMethod.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.ByParams.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.ByType.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.OneBaseline.Default.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.ByCategory.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.ByJob.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.ByMethod.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.ByParams.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.ByType.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselines.Default.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.ByCategory.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.ByJob.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.ByMethod.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.ByParams.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.ByType.approved.txt
… tests/BenchmarkDotNet.Tests/Exporters/ApprovedFiles/MarkdownExporterMultipleClassApprovalTests.LogicalRuleTest.TwoBaselinesWithJobs.Default.approved.txt