Fixes #1164
Fixes #2022

Look at "Update existing tests" and "Update new tests" to see what has changed.

# Fixes:

## https://github.com/dotnet/BenchmarkDotNet/issues/1164#issue-448217544
<details>

Master:
```
|       Type | Method |     Mean |    Error |   StdDev | Ratio |
|----------- |------- |---------:|---------:|---------:|------:|
| Formatting |    Old | 15.52 ms | 0.306 ms | 0.017 ms |  1.00 |
|    Parsing |    Old | 15.54 ms | 1.847 ms | 0.101 ms |  1.00 |
| Formatting |    New | 15.48 ms | 2.415 ms | 0.132 ms |  1.00 |
|    Parsing |    New | 15.49 ms | 1.244 ms | 0.068 ms |  1.00 |

```
PR:
```
|       Type | Method |     Mean |    Error |   StdDev | Ratio |
|----------- |------- |---------:|---------:|---------:|------:|
| Formatting |    Old | 15.46 ms | 0.039 ms | 0.035 ms |  1.00 |
| Formatting |    New | 15.62 ms | 0.081 ms | 0.075 ms |  1.01 |
|            |        |          |          |          |       |
|    Parsing |    Old | 15.59 ms | 0.039 ms | 0.036 ms |  1.00 |
|    Parsing |    New | 15.54 ms | 0.076 ms | 0.071 ms |  1.00 |
```
</details>

## https://github.com/dotnet/BenchmarkDotNet/issues/2022#issue-1279202717

<details>

Master:
```
         Type |       Method | Pending |     Mean |   Error |  StdDev | LogicalGroup |
------------- |------------- |-------- |---------:|--------:|--------:|------------- |
   AsyncAwait |     Callback |   False | 102.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith |     Callback |   False | 102.0 ns | 6.09 ns | 1.58 ns |            * |
   AsyncAwait | ProtoPromise |   False | 202.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith | ProtoPromise |   False | 202.0 ns | 6.09 ns | 1.58 ns |            * |
   AsyncAwait |   RsgPromise |   False | 302.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith |   RsgPromise |   False | 302.0 ns | 6.09 ns | 1.58 ns |            * |
   AsyncAwait |         Task |   False | 402.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith |         Task |   False | 402.0 ns | 6.09 ns | 1.58 ns |            * |
   AsyncAwait |      UniTask |   False | 502.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith |      UniTask |   False | 502.0 ns | 6.09 ns | 1.58 ns |            * |
   AsyncAwait |    ValueTask |   False | 602.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWith |    ValueTask |   False | 602.0 ns | 6.09 ns | 1.58 ns |            * |
```

PR
```
         Type |       Method | Pending |     Mean |   Error |  StdDev | LogicalGroup |
------------- |------------- |-------- |---------:|--------:|--------:|------------- |
   AsyncAwait |     Callback |   False | 102.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
   AsyncAwait | ProtoPromise |   False | 202.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
   AsyncAwait |   RsgPromise |   False | 302.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
   AsyncAwait |         Task |   False | 402.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
   AsyncAwait |      UniTask |   False | 502.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
   AsyncAwait |    ValueTask |   False | 602.0 ns | 6.09 ns | 1.58 ns |   AsyncAwait |
              |              |         |          |         |         |              |
 ContinueWith |     Callback |   False | 102.0 ns | 6.09 ns | 1.58 ns | ContinueWith |
 ContinueWith | ProtoPromise |   False | 202.0 ns | 6.09 ns | 1.58 ns | ContinueWith |
 ContinueWith |   RsgPromise |   False | 302.0 ns | 6.09 ns | 1.58 ns | ContinueWith |
 ContinueWith |         Task |   False | 402.0 ns | 6.09 ns | 1.58 ns | ContinueWith |
 ContinueWith |      UniTask |   False | 502.0 ns | 6.09 ns | 1.58 ns | ContinueWith |
 ContinueWith |    ValueTask |   False | 602.0 ns | 6.09 ns | 1.58 ns | ContinueWith |

```
</details>

## https://github.com/dotnet/BenchmarkDotNet/issues/1864#issuecomment-1001346715
<details>

Master
```
                  Type |          Method |   N |     Mean |   Error |  StdDev | LogicalGroup |
---------------------- |---------------- |---- |---------:|--------:|--------:|------------- |
          AsyncPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
         AsyncResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
          AwaitPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
         AwaitResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWithFromValue | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
   ContinueWithPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
  ContinueWithResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |            * |
          AsyncPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
         AsyncResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
          AwaitPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
         AwaitResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
 ContinueWithFromValue | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
   ContinueWithPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |
  ContinueWithResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |            * |

```
PR
```
                  Type |          Method |   N |     Mean |   Error |  StdDev |          LogicalGroup |
---------------------- |---------------- |---- |---------:|--------:|--------:|---------------------- |
          AsyncPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |          AsyncPending |
          AsyncPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |          AsyncPending |
                       |                 |     |          |         |         |                       |
         AsyncResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |         AsyncResolved |
         AsyncResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |         AsyncResolved |
                       |                 |     |          |         |         |                       |
          AwaitPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |          AwaitPending |
          AwaitPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |          AwaitPending |
                       |                 |     |          |         |         |                       |
         AwaitResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |         AwaitResolved |
         AwaitResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |         AwaitResolved |
                       |                 |     |          |         |         |                       |
 ContinueWithFromValue | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns | ContinueWithFromValue |
 ContinueWithFromValue | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns | ContinueWithFromValue |
                       |                 |     |          |         |         |                       |
   ContinueWithPending | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |   ContinueWithPending |
   ContinueWithPending | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |   ContinueWithPending |
                       |                 |     |          |         |         |                       |
  ContinueWithResolved | ProtoPromise_V1 | 100 | 102.0 ns | 6.09 ns | 1.58 ns |  ContinueWithResolved |
  ContinueWithResolved | ProtoPromise_V2 | 100 | 202.0 ns | 6.09 ns | 1.58 ns |  ContinueWithResolved |
```
</details>

## https://github.com/dotnet/BenchmarkDotNet/issues/1864#issuecomment-1031364986
<details>

Master
```
                 Type |           Method |     Mean |   Error |  StdDev | LogicalGroup |
--------------------- |----------------- |---------:|--------:|--------:|------------- |
      ArrayBenchmarks |            Empty | 102.0 ns | 6.09 ns | 1.58 ns |            * |
      EmptyBenchmarks |  EnumerableEmpty | 102.0 ns | 6.09 ns | 1.58 ns |            * |
 EnumerableBenchmarks |            Empty | 102.0 ns | 6.09 ns | 1.58 ns |            * |
     StringBenchmarks | ConstructorArray | 102.0 ns | 6.09 ns | 1.58 ns |            * |
      EmptyBenchmarks |       ArrayEmpty | 202.0 ns | 6.09 ns | 1.58 ns |            * |
     StringBenchmarks |  ConstructorSpan | 202.0 ns | 6.09 ns | 1.58 ns |            * |

```
PR
```
                 Type |           Method |     Mean |   Error |  StdDev |         LogicalGroup |
--------------------- |----------------- |---------:|--------:|--------:|--------------------- |
      ArrayBenchmarks |            Empty | 102.0 ns | 6.09 ns | 1.58 ns |      ArrayBenchmarks |
                      |                  |          |         |         |                      |
      EmptyBenchmarks |  EnumerableEmpty | 102.0 ns | 6.09 ns | 1.58 ns |      EmptyBenchmarks |
      EmptyBenchmarks |       ArrayEmpty | 202.0 ns | 6.09 ns | 1.58 ns |      EmptyBenchmarks |
                      |                  |          |         |         |                      |
 EnumerableBenchmarks |            Empty | 102.0 ns | 6.09 ns | 1.58 ns | EnumerableBenchmarks |
                      |                  |          |         |         |                      |
     StringBenchmarks | ConstructorArray | 102.0 ns | 6.09 ns | 1.58 ns |     StringBenchmarks |
     StringBenchmarks |  ConstructorSpan | 202.0 ns | 6.09 ns | 1.58 ns |     StringBenchmarks |
```
</details>


# Tests
Yes, that's a lot of tests, but now it's easy to change the ordering logic as there are a lot of cases covered.

For example, this change produces a better output, but it breaks output for `OneBaseline` in some test (it starts display baseline numbers instead a question mark):
```diff
- if (hasMultipleTypes)
+ if (hasMultipleTypes && explicitRules.IsEmpty())
```

# Question
What is the correct table for this?

<details>

```C#
[GroupBenchmarksBy(BenchmarkLogicalGroupRule.ByMethod)]
[LogicalGroupColumn, CategoriesColumn, BaselineColumn]
[BenchmarkCategory("A")]
[SimpleJob(id: "Job1"), SimpleJob(id: "Job2")]
public class Bench1
{
    [Params(10, 20)] public int Param;
    [Benchmark(Baseline = true)] public void Foo() { }
    [Benchmark] public void Bar() { }
}

[GroupBenchmarksBy(BenchmarkLogicalGroupRule.ByMethod)]
[LogicalGroupColumn, CategoriesColumn, BaselineColumn]
[BenchmarkCategory("B")]
public class Bench2
{
    [Params(10, 20)] public int Param;
    [Benchmark(Baseline = true)] public void Foo() { }
    [Benchmark] public void Bar() { }
}
```

Master/PR output is:
```
   Type | Method |        Job | Categories | Param |     Mean |   Error |  StdDev | Ratio | RatioSD |                     LogicalGroup | Baseline |
------- |------- |----------- |----------- |------ |---------:|--------:|--------:|------:|--------:|--------------------------------- |--------- |
 Bench1 |    Foo |       Job1 |          A |    10 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |       Foo-Bench1-[Param=10]-Job1 |      Yes | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Foo |       Job2 |          A |    10 | 302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |       Foo-Bench1-[Param=10]-Job2 |      Yes |
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Foo |       Job1 |          A |    20 | 502.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |       Foo-Bench1-[Param=20]-Job1 |      Yes | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Foo |       Job2 |          A |    20 | 702.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 |       Foo-Bench1-[Param=20]-Job2 |      Yes |
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench2 |    Foo | DefaultJob |          B |    10 | 102.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 | Foo-Bench2-[Param=10]-DefaultJob |      Yes | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench2 |    Foo | DefaultJob |          B |    20 | 302.0 ns | 6.09 ns | 1.58 ns |  1.00 |    0.00 | Foo-Bench2-[Param=20]-DefaultJob |      Yes | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Bar |       Job1 |          A |    10 | 202.0 ns | 6.09 ns | 1.58 ns |     ? |       ? |       Bar-Bench1-[Param=10]-Job1 |       No | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Bar |       Job2 |          A |    10 | 402.0 ns | 6.09 ns | 1.58 ns |     ? |       ? |       Bar-Bench1-[Param=10]-Job2 |       No |
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Bar |       Job1 |          A |    20 | 602.0 ns | 6.09 ns | 1.58 ns |     ? |       ? |       Bar-Bench1-[Param=20]-Job1 |       No | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench1 |    Bar |       Job2 |          A |    20 | 802.0 ns | 6.09 ns | 1.58 ns |     ? |       ? |       Bar-Bench1-[Param=20]-Job2 |       No |
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench2 |    Bar | DefaultJob |          B |    10 | 202.0 ns | 6.09 ns | 1.58 ns |     ? |       ? | Bar-Bench2-[Param=10]-DefaultJob |       No | ^
        |        |            |            |       |          |         |         |       |         |                                  |          |
 Bench2 |    Bar | DefaultJob |          B |    20 | 402.0 ns | 6.09 ns | 1.58 ns |     ? |       ? | Bar-Bench2-[Param=20]-DefaultJob |       No | ^

```
</details>