@timcassell 2022-10-18
A follow-up to this may want to add a `ByRuntime` group rule, which by default should be ordered before `ByType`.