@YegorStepanov 2022-10-18
I don't like that grouping `ByMethod` groups them into separate groups, but this seems to be the correct behavior:

#### How the grouping works:

1) Look at the `LogicalGroup` column. It shows the keys to sort by (the keys are truncated from the right if ALL keys in the table are the same)

2) There is default grouping order: `ByType, ByCategory, ByParams, ByJob, ByMethod`

3) When we add a grouping manually (`[GroupBenchmarksBy(BenchmarkLogicalGroupRule.ByMethod)]`) it adds this key to the left. So it turns out that the groups consist of 1 element.