@YegorStepanov 2022-10-18
I'm converting it to draft until I figure out why nothing changes due to swapping `ByCategory`<->`ByType`