@YegorStepanov 2022-10-21
I made the most conservative mode. I don't like that there were more groups than I expected (look at the last commit)

assign to @AndreyAkinshin